[WEB] transpose the matrix on construction
design is as follows:

you pass in an array that is row-major, and we transpose it internally
in the matrix constructor, but you don't have to think about it if you
use the provided accessors. The only time that you need to care about
the internal storage is when you use the matrix.Float32Array accessor,
which returns the transposed, column-major array. This is fine, because
generally when you use matrix.Float32Array accessor, you're passing it
into some WebGL function, which expects a column-major flat matrix